test: steps 분리
