+++ src/test/java/subway/acceptance/auth/AuthSteps.java
@@ -17,4 +17,6 @@ public class AuthSteps {
                 .then().log().all()
                 .extract();
     }
+
+
 }
