+++ src/test/java/subway/acceptance/member/MemberSteps.java
@@ -3,7 +3,9 @@ package subway.acceptance.member;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
+import subway.acceptance.auth.AuthFixture;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -63,5 +65,33 @@ public class MemberSteps {
         assertThat(response.jsonPath().getString("id")).isNotNull();
         assertThat(response.jsonPath().getString("email")).isEqualTo(email);
         assertThat(response.jsonPath().getInt("age")).isEqualTo(age);
+    }
+
+    public static ExtractableResponse<Response> 회원_조회_API(String memberLocation, String accessToken) {
+        var response = RestAssured.given().log().all()
+                .header("Authorization", AuthFixture.BEARER_만들기(accessToken))
+                .when().get(memberLocation)
+                .then().log().all()
+                .extract();
+        return response;
+    }
+
+    public static ExtractableResponse<Response> 회원_삭제_API(String memberLocation, String accessToken) {
+        return RestAssured.given().log().all()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .header("Authorization", AuthFixture.BEARER_만들기(accessToken))
+                .when().delete(memberLocation)
+                .then().log().all()
+                .statusCode(HttpStatus.NO_CONTENT.value()).extract();
+    }
+
+    public static ExtractableResponse<Response> 회원_수정_API(String memberLocation, String accessToken, Map<String, String> params) {
+        return RestAssured.given().log().all()
+                .body(params)
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .header("Authorization", AuthFixture.BEARER_만들기(accessToken))
+                .when().put(memberLocation)
+                .then().log().all()
+                .statusCode(HttpStatus.OK.value()).extract();
     }
 }
